Replace Kube Java callback with Rx Mono
