+++ titus-common-runtime/src/main/java/com/netflix/titus/runtime/connector/kubernetes/DefaultKubeApiFacade.java
@@ -29,7 +29,6 @@ import com.netflix.titus.runtime.connector.kubernetes.v1.V1OpportunisticResource
 import io.kubernetes.client.common.KubernetesObject;
 import io.kubernetes.client.informer.SharedIndexInformer;
 import io.kubernetes.client.informer.SharedInformerFactory;
-import io.kubernetes.client.openapi.ApiCallback;
 import io.kubernetes.client.openapi.ApiClient;
 import io.kubernetes.client.openapi.ApiException;
 import io.kubernetes.client.openapi.apis.CoreV1Api;
@@ -46,18 +45,17 @@ import io.kubernetes.client.util.CallGeneratorParams;
 import okhttp3.Call;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import reactor.core.publisher.Mono;
 
 import static com.netflix.titus.runtime.connector.kubernetes.KubeApiClients.createSharedInformerFactory;
 
 @Singleton
 public class DefaultKubeApiFacade implements KubeApiFacade {
 
+    private static final Logger logger = LoggerFactory.getLogger(DefaultKubeApiFacade.class);
+
     public static final String FAILED = "Failed";
     public static final String BACKGROUND = "Background";
-
-    public static final String DEFAULT_NAMESPACE = "default";
-
-    private static final Logger logger = LoggerFactory.getLogger(DefaultKubeApiFacade.class);
 
     private static final String KUBERNETES_NAMESPACE = "default";
 
@@ -136,12 +134,14 @@ public class DefaultKubeApiFacade implements KubeApiFacade {
 
     @Override
     public void createNamespacedPod(String namespace, V1Pod pod) throws ApiException {
-        coreV1Api.createNamespacedPod(DEFAULT_NAMESPACE, pod, null, null, null);
+        coreV1Api.createNamespacedPod(namespace, pod, null, null, null);
     }
 
     @Override
-    public Call createNamespacedPodAsync(String namespace, V1Pod pod, ApiCallback<V1Pod> callback) throws ApiException {
-        return coreV1Api.createNamespacedPodAsync(namespace, pod, null, null, null, callback);
+    public Mono<V1Pod> createNamespacedPodAsync(String namespace, V1Pod pod) {
+        return KubeUtil.toReact(handler ->
+                coreV1Api.createNamespacedPodAsync(namespace, pod, null, null, null, handler)
+        );
     }
 
     @Override
