+++ titus-common-runtime/src/main/java/com/netflix/titus/runtime/connector/kubernetes/KubeUtil.java
@@ -16,9 +16,24 @@
 
 package com.netflix.titus.runtime.connector.kubernetes;
 
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+
+import io.kubernetes.client.openapi.ApiCallback;
+import io.kubernetes.client.openapi.ApiException;
 import io.kubernetes.client.openapi.models.V1ObjectMeta;
+import okhttp3.Call;
+import reactor.core.publisher.Mono;
 
 public final class KubeUtil {
+
+    /**
+     * Like {@link Function}, but with {@link ApiException} throws clause.
+     */
+    public interface KubeFunction<I, O> {
+        O apply(I argument) throws ApiException;
+    }
 
     /**
      * Get Kube object name
@@ -29,5 +44,41 @@ public final class KubeUtil {
         }
 
         return metadata.getName();
+    }
+
+    public static <T> Mono<T> toReact(KubeFunction<ApiCallback<T>, Call> handler) {
+        return Mono.create(sink -> {
+            Call call;
+            try {
+                call = handler.apply(new ApiCallback<T>() {
+                    @Override
+                    public void onFailure(ApiException e, int statusCode, Map<String, List<String>> responseHeaders) {
+                        sink.error(e);
+                    }
+
+                    @Override
+                    public void onSuccess(T result, int statusCode, Map<String, List<String>> responseHeaders) {
+                        if (result == null) {
+                            sink.success();
+                        } else {
+                            sink.success(result);
+                        }
+                    }
+
+                    @Override
+                    public void onUploadProgress(long bytesWritten, long contentLength, boolean done) {
+                    }
+
+                    @Override
+                    public void onDownloadProgress(long bytesRead, long contentLength, boolean done) {
+                    }
+                });
+            } catch (ApiException e) {
+                sink.error(e);
+                return;
+            }
+
+            sink.onCancel(call::cancel);
+        });
     }
 }
